mem-cache: Remove power-of-2 requirement for TreePLRU num leaves
Remove the requirement in TreePLRU's implementation that the number of
leaves (i.e., the number of cache ways) be a power of two. Firstly, on
some recent processors, this is not the case---for example, Intel Golden
Cove's L1D has 12 ways.) Secondly, The implementation of TreePLRU
appears to work just fine as-is with a way count that's not a power of
two.

Change-Id: If2a27dc5bbe7a8e96684f79ce791df5c0b582230